Deleted lasers, tore down spellcard and rebuilt it (bare bones)
Shit to this fuck I hate lasers. I realised that the old spellcard method had become so outdated that everything in it that wasnt the effects were broken. I hate my shit now I have to redo it anyways this is todays commit. Wait what language do I write the commits in?